fix: reset on new day -  extracted the commands in their own function called initExtensionCommands -  extracted the logic of reset on a new day in a function called isNewDayHandler - in that function,  we update the global state data then update the dailyData and lastServerSync if it is a new day
