+++ apps/vscode-extension/src/extension.ts
@@ -3,12 +3,8 @@ import { filesData, languagesData } from "./constants";
 import addStatusBarItem from "./utils/addStatusBarItem";
 import calculateTime from "./utils/calculateTime";
 import fetchInitialData from "./utils/fetchInitialData";
-import getGlobalStateData from "./utils/getGlobalStateData";
-import login from "./utils/auth/login";
-import logout from "./utils/auth/logout";
-import openDashBoard from "./utils/openDashBoard";
+import initExtensionCommands from "./utils/initExtensionCommands";
 import periodicSyncData from "./utils/periodicSyncData";
-import register from "./utils/auth/register";
 import setStatusBarItem from "./utils/setStatusBarItem";
 
 let extensionContext: vscode.ExtensionContext;
@@ -28,7 +24,6 @@ export async function activate(context: vscode.ExtensionContext) {
   setStatusBarItem(timeSpent, statusBarItem);
 
   // initialize the time for each language found
-
   Object.keys(initialLanguagesData).forEach((languageName) => {
     const timeSpent = initialLanguagesData[languageName];
     const now = performance.now();
@@ -67,86 +62,10 @@ export async function activate(context: vscode.ExtensionContext) {
     await periodicSyncData(context, statusBarItem, getTime);
   }, 60000);
 
-  // debugging commands
-  const showCurrentLanguagesDataCommand = vscode.commands.registerCommand(
-    "MoonCode.showCurrentLanguagesData",
-    () => {
-      const { languagesData } = getTime();
-      vscode.window.showInformationMessage(
-        `currentLanguagesData: ${JSON.stringify(Object.entries(languagesData).map(([key, { elapsedTime }]) => `${key}: ${elapsedTime} seconds`))}`,
-      );
-    },
-  );
-
-  const showInitialLanguagesDataCommand = vscode.commands.registerCommand(
-    "MoonCode.showInitialLanguagesData",
-    async () => {
-      const globalStateData = await getGlobalStateData();
-
-      vscode.window.showInformationMessage(
-        `initialLanguagesData from server: ${JSON.stringify(Object.entries(initialLanguagesData).map(([key, elapsedTime]) => `${key}: ${elapsedTime} seconds`))}`,
-      );
-
-      vscode.window.showInformationMessage(
-        `Global state content: ${JSON.stringify(globalStateData)}`,
-      );
-    },
-  );
-
-  const showCurrentFilesDataCommand = vscode.commands.registerCommand(
-    "MoonCode.showCurrentFilesData",
-    () => {
-      const { filesData: currentFilesData } = getTime();
-      vscode.window.showInformationMessage(
-        `currentFilesData: ${JSON.stringify(Object.entries(currentFilesData).map(([key, { elapsedTime }]) => `${key}: ${elapsedTime} seconds`))}`,
-      );
-    },
-  );
-
-  const showInitialFilesDataCommand = vscode.commands.registerCommand(
-    "MoonCode.showInitialFilesData",
-    async () => {
-      vscode.window.showInformationMessage(
-        `initialFilesData from server: ${JSON.stringify(Object.entries(initialFilesData).map(([key, { timeSpent: elapsedTime }]) => `${key}: ${elapsedTime} seconds`))}`,
-      );
-    },
-  );
-
-  const loginCommand = vscode.commands.registerCommand(
-    "MoonCode.login",
-    async () => {
-      await login();
-    },
-  );
-
-  const registerCommand = vscode.commands.registerCommand(
-    "MoonCode.register",
-    async () => {
-      await register();
-    },
-  );
-
-  const logoutCommand = vscode.commands.registerCommand(
-    "MoonCode.logout",
-    async () => {
-      await logout();
-    },
-  );
-
-  const openDashBoardCommand = vscode.commands.registerCommand(
-    "MoonCode.openDashBoard",
-    openDashBoard,
-  );
-
-  context.subscriptions.push(
-    showInitialLanguagesDataCommand,
-    showCurrentLanguagesDataCommand,
-    showInitialFilesDataCommand,
-    showCurrentFilesDataCommand,
-    loginCommand,
-    registerCommand,
-    logoutCommand,
-    openDashBoardCommand,
+  initExtensionCommands(
+    getTime,
+    initialLanguagesData,
+    initialFilesData,
     statusBarItem,
   );
 }
